+++ modules/core/02-client/types/events.go
@@ -15,6 +15,8 @@ const (
 	AttributeKeyHeader            = "header"
 	AttributeKeyUpgradePlanTitle  = "title"
 	AttributeKeyUpgradePlanHeight = "height"
+	AttributeKeyUpgradeStore      = "upgrade_store"
+	AttributeKeyUpgradePlanHeight = "upgrade_plan_height"
 )
 
 // IBC client events vars
@@ -25,6 +27,7 @@ var (
 	EventTypeSubmitMisbehaviour    = "client_misbehaviour"
 	EventTypeUpdateClientProposal  = "update_client_proposal"
 	EventTypeUpgradeClientProposal = "upgrade_client_proposal"
+	EventTypeUpgradeChain          = "upgrade_chain"
 
 	AttributeValueCategory = fmt.Sprintf("%s_%s", host.ModuleName, SubModuleName)
 )